More Refactoring
Also cleaned up the really gross options classes, literally pouring
performance in the trash using them(Loading them idk but using is just
aweful) If they stopped working id assume so to happen with the XML
handlers Damian has... Ill give them a new handler if that is the case.
Im just save stating large collections of refactoring - im not worried
if small things broke atm.

+++ Source/Bindings/Legacy/Constants.vb
@@ -18,8 +18,8 @@
     Friend Const MAX_SKILLS As Byte = 255
     Friend Const MAX_TRADES As Byte = 20
     Friend Const NAME_LENGTH As Byte = 20
-    Friend Const MAX_PARTIES As Long = 35
-    Friend Const MAX_PARTY_MEMBERS As Long = 4
+    Friend Const MAX_PARTIES As Integer = 35
+    Friend Const MAX_PARTY_MEMBERS As Integer = 4
     Friend Const MAX_RANDOM_TYPES As Byte = 16
     Friend Const MAX_PETS As Integer = 100
     Friend Const MAX_DOTS As Integer = 30
